Don't store string slices on the token
In most of the tokens, all slices are the same

Example:
Collon will always be ":"
And While will always be "while"

It makes no sense to store the strings.

Instead, store the slice of the source string the token
represents, and convert it into an owned value in the cases
in which it is needed.