error_manager: Allow downcasting of Error, to a concrete type
Make Error trait bounded to Any. This allows us to upcast any
`&dyn Error` to a `&dyn Any`, and then downcast that to a concrete
error type.

This is useful when we are testing a compilation stage and we know that
all the `dyn Error` stored inside the ErrorManager will be insances of
a concrete type

For example, when testing the identification stage, downcasting Error
to IdentificationError allows us to inspect the IdentificationError's
fields, and assert that an input program emits the expected errors.