Replace session with interner
Having all the global state behind a single `Session` structure can be
bad in the future. If we had more fields inside it, we would need to
lock the whole session to access a single field.

It's best to keep the interner inside it's own crate.

Also, remove those "with_session_interner" functions. It's cleaner if we
hide that stuff from the outside. The interner crate only exports a
`Symbol` struct that provides a way to intern and resolve from and to
strings.