Start working to support multiple source files
Introduce new structs `Source` and `SourceFile`, that represent...
well, a source file.
The Span now points to a specific source file.
Use that new information to print the filename when emiting errors.
To avoid growing the size of Span, change the representation of the
offset and len from `usize` to `u32`. u32 seems like a reasonable type
for the offset on a file.