span: Remove fileid
Instead of having a fileid for each span, use a base offset for each
span.

When adding a new file to the source map, we track the base offset for
that file.

This is
File1: offset = 0, len = 120
File2: offset = 120, len = 80
File3: offset = 200, len = 20

When building a span for File1 with offset 10, we add the base offset
of that file (120). So we'll have a span with offset 130.

To retreive the slice of a Span with offset 130, we search for the
file that contains that offset (File2), and we substract the base offset
from the span. (130 - 120 = 10)